requirement.txt
fastapi
mcp
python-dotenv
pydantic
openai-agents
asyncio